support for fractional gSig

+++ caiman/source_extraction/cnmf/initialization.py
@@ -822,24 +822,25 @@ def imblur(Y, sig=5, siz=11, nDimBlur=None, kernel=None, opencv=True):
             siz = siz * np.ones(nDimBlur)
 
         X = Y.copy()
+
         if opencv and nDimBlur == 2:
             if X.ndim > 2:
                 # if we are on a video we repeat for each frame
                 for frame in range(X.shape[-1]):
                     if sys.version_info >= (3, 0):
                         X[:, :, frame] = cv2.GaussianBlur(X[:, :, frame], tuple(
-                            siz), sig[0], None, sig[1], cv2.BORDER_CONSTANT)
+                            siz.astype(np.int)), sig[0], None, sig[1], cv2.BORDER_CONSTANT)
                     else:
-                        X[:, :, frame] = cv2.GaussianBlur(X[:, :, frame], tuple(siz), sig[
+                        X[:, :, frame] = cv2.GaussianBlur(X[:, :, frame], tuple(siz.astype(np.int)), sig[
                                                           0], sig[1], cv2.BORDER_CONSTANT, 0)
 
             else:
                 if sys.version_info >= (3, 0):
                     X = cv2.GaussianBlur(
-                        X, tuple(siz), sig[0], None, sig[1], cv2.BORDER_CONSTANT)
+                        X, tuple(siz.astype(np.int)), sig[0], None, sig[1], cv2.BORDER_CONSTANT)
                 else:
                     X = cv2.GaussianBlur(
-                        X, tuple(siz), sig[0], sig[1], cv2.BORDER_CONSTANT, 0)
+                        X, tuple(siz.astype(np.int)), sig[0], sig[1], cv2.BORDER_CONSTANT, 0)
         else:
             for i in range(nDimBlur):
                 h = np.exp(
